Examples are now cleared on form submit.
Eren discovered that examples in fields are indistinguishable from actual
values when a form is submitted. This means that fields that have not been
filled in by a user will use its example as its value when it should instead
be blank. This is fixed by simply clearing fields that have the example class
on form submit.

I use the submit event of the parent form rather than the click event of any
submit button as it seems more correct (after all, a user might not click a
submit button at all to submit a form).

In light of reading up on the use of JavaScript in production, I now minimise
the file using the YUICompressor which compresses better with GZip than
Dean Edward's Packer. Seemingly contrary to this, however, I have replaced
the minimised version of jQuery with the packed one just to reduce disk
space usage as the file is only for testing and is not really part of the
bundle.