Allows for values the same as example text and only fills empty fields on load.
Improved the conditions when inserting example text and when clearing fields for
entry. Instead of just checking whether the value of a field matches the example
text given on focus, the class of the field is also checked to more reliably
determine whether it is an example or not.

This means that example texts are now valid possible values whereas previously such
values would have been cleared. e.g. If the example for a field is "Bob Smith",
putting "Bob Smith" as the actual value will now work.

In the same vein, examples are only set on page load if the selected fields
are empty. Previously, examples would always overwrite existing field values.