Changed the header comment of the script to contain more information.
Changed the LICENSE file to have a .txt extension as it would bizarrely become
executable when extracted from the zip file. Added a reference to the license
in the comment header which has also been added to the compressed version
(as that should be the most used version after all).