Fixed the caching bug on Gecko-based browsers by clearing values on unload.
While everything previously worked fine on Safari, there were problems with
Gecko-based browsers such as Firefox and Camino whereby the browser would
cache the example values. This meant that a simple refresh of the page would
convert example values into *real* values thereby breaking the plugin entirely.

To overcome this, I now bind an event on page unload to clear all the example
values so by the time that a Gecko browser comes to cache the page, those
fields are all blank.

I also removed support for the Metadata plugin (as I didn't use it and only
added it based on an article about jQuery plugins) and reduced the number
of duplicate event bindings.